kmod: break ranges of map pages handling
When `->map_pages()` is called, it receives a range around the actual
page fault as an optmimistic optmization to pre-load pages that are
likely to fault as well.

This patch updates the custom page fault handler to break this range
into chunks where we don't need any custom logic and the ones that do.