kmod: restore base VMA vm_ops on cleanup
When the kernel module cleans up a memory overlay, it must restore the
base VMA origin `vm_ops` because the memory allocated for the highjacked
`vm_ops` is cleared.

Without restore it, any operations on the VMA after the cleanup routine
runs may result in the error `kernel tried to execute NX-protected
page`.