kmod: don't lock on hashtable cleanup
The hashtable cleanup routine runs when the kernel module is unloaded,
so it's not possible to know if the process that owns the memory
overlays are still running. If they're not, the kernel module is not
able to acquire a `mmap_write_lock()` on the MM.

By not locking, we assume users coordinates the kernel module shutdown
by first stopping and cleaning up any userspace program that uses memory
overlays. Any data leftover is assumed to be from programs that crashed
and so can be cleaned up without a lock.