kmod: remove the use of the VMA's private data
Since page fault are triggered without direct control from the userspace
program, it's not possible to pass the generated UUID back to the kernel
module to find the appropriate memory overlay request data.

To overcome this problem, the kernel module would store the UUID in the
VMA's `->vm_private_data` field so it could read it back when handling a
page fault. But this field can also be used by other kernel modules and
drivers, so it's best to avoid relying on it.

This commit modifies the hashtable key to use the VMA virtual space.
Since VMAs are stored in kernel space, their virtual address are unique
for running processes.

The virtual address can be reused in cases where the userspace program
exits without properly cleaning up its resources in the kernel module
(such as after an unexpected crash).

In these cases, the new VMA has not been hijacked yet, so the kernel
module deletes the leftover stale data and proceeds with the rest of the
logic.

If the VMA has been hijacked, it's likely that the userspace program
called the kernel module twice for the same VMA, so an error is returned
as it is expected that the program cleans up the previous entry before
registering a new one.