kmod: fix bug on segments cleanup
`xa_store_range()` may break the range into multiple entries depending
on the range size. When this happens, iterating over the entries with
`xa_for_each()` returns the same segment multiple times.

In the cleanup routine this caused `kvfree()` to be called on segments
multiple times and the kernel to panic.

To prevent this, the iteration index needs to be moved past the segment
range, so the next element returned is a new segment.

In order to test this scenario, the `page_fault.c` test was refactored
to call `IOCTL_MEM_OVERLAY_CLEANUP_CMD` after each test case.