kmod: fix page fault range split logic
When handling a page fault with `->map_pages()`, the Linux kernel
provides a range of pages to handle around the actual fault. The
hijacked handler splits this range into sub-ranges that are not part of
an overlay and those that are.

If the range overlaps with an overlay, it may have some non-overlap
sub-range before it.

```
Page fault range overlaps with an overlay segment.

start_pgoff                                 end_pgoff
  start   seg->start_pgoff     seg->end_pgoff  end
    |           |                   |           |
  +---+---+---+---+---+---+---+---+---+---+---+---+
  |   |   |   | x | x | x | x | x | x |   |   |   |
  +---+---+---+---+---+---+---+---+---+---+---+---+

Handle non-overlap range before the overlay.

          seg->start_pgoff
start_pgoff     |                           end_pgoff
  start    end  |              seg->end_pgoff   |
    |       |   |                   |           |
  +---+---+---+---+---+---+---+---+---+---+---+---+
  |   |   |   | x | x | x | x | x | x |   |   |   |
  +---+---+---+---+---+---+---+---+---+---+---+---+

Handle overlapping overlay range.

          seg->start_pgoff     seg->end_pgoff
start_pgoff   start                end      end_pgoff
    |           |                   |           |
  +---+---+---+---+---+---+---+---+---+---+---+---+
  |   |   |   | x | x | x | x | x | x |   |   |   |
  +---+---+---+---+---+---+---+---+---+---+---+---+
```

But if the fault range starts in the middle of an overlay segment, the
handler should not handle the sub-range starting at the segment start,
but rather keep the start where it is.

```
Page fault range starts after an overlay segment.

seg->start_pgoff     seg->end_pgoff
      |   start_pgoff     |   end_pgoff
      |     start         |      end
      |       |           |       |
    +---+---+---+---+---+---+---+---+---+---+---+---+
    | x | x | x | x | x | x |   |   |   |   |   |   |
    +---+---+---+---+---+---+---+---+---+---+---+---+
```

So the start should only change if there's a non-overlay sub-range
before the overlay area.

For some reason this bug was only reliably reproduced with the
`page_fault_benchmark` test with `N=10`. Refactor the benchmark code to
help parameterized it.

The benchmark code was also affected by OS caching, so clear it before
running each test.